wip: need to fix unsend

+++ src/utils/commands/saveSendLogs.ts
@@ -1,17 +1,13 @@
-import { DiscordEvent } from '../types/types';
+import { DiscordEvent, EventLog } from '../schema';
 import { assert } from '../assert';
 import fs from 'fs';
 import path from 'path';
-import readline from 'readline';
+// import readline from 'readline';
 
 async function appendLinesToFile(filePath: string, lines: string[]): Promise<void> {
     try {
-        if (!fs.existsSync(filePath)) {
-            await fs.promises.writeFile(filePath, '', 'utf8');
-        }
-
+        if (!fs.existsSync(filePath)) await fs.promises.writeFile(filePath, '', 'utf8');
         const data = lines.join('\n') + '\n';
-
         await fs.promises.appendFile(filePath, data, 'utf8');
     } catch (error) {
         assert(error instanceof Error, 'Error should be an instance of Error');
@@ -21,49 +17,20 @@ async function appendLinesToFile(filePath: string, lines: string[]): Promise<voi
     }
 }
 
-async function readLastNLines(filePath: string, n: number): Promise<string[]> {
-    const fileStream = fs.createReadStream(filePath, { encoding: 'utf8' });
-    const rl = readline.createInterface({
-        // FIXME: The type definition here is incorrect, force-case for now
-        input: fileStream as unknown as NodeJS.ReadableStream,
-        crlfDelay: Infinity,
-    });
-
-    const buffer: string[] = [];
-
-    // Read the file line by line
-    for await (const line of rl) {
-        buffer.push(line);
-        if (buffer.length > n) {
-            buffer.shift();
-        }
-    }
-
-    rl.close();
-
-    return buffer;
-}
-
-type EventLog = {
-    subject: string;
-    summary: string;
-    calendarId: string;
-    messageId: string;
-    threadId: string;
-    guildId: string;
-};
-
-function formatDiscordEvent(event: DiscordEvent): string[] {
+function formatDiscordEvent(event: DiscordEvent, timestamp: Date): string[] {
     const { events, messageId, threadId, guildId } = event;
-    const eventLogs: EventLog[] = events.map((event) => {
+    const eventLogs = events.map((event) => {
         const { calendarId, subject, summary } = event;
-        return { subject, summary, calendarId, messageId, threadId, guildId };
-    });
+        return { timestamp, subject, summary, calendarId, messageId, threadId, guildId };
+    }) as EventLog[];
 
     return eventLogs.map((event) => JSON.stringify(event));
 }
+
 export async function saveSendLogs(successfulEvents: DiscordEvent[]) {
-    const logs: string[][] = successfulEvents.map(formatDiscordEvent);
+    const timestamp = new Date();
+    const logs: string[][] = successfulEvents.map((event) => formatDiscordEvent(event, timestamp));
     const logFilePath = path.join(__dirname, '../../data/sendLogs.txt');
     await appendLinesToFile(logFilePath, logs.flat());
+    await appendLinesToFile(logFilePath, ['[ batch dispatch ]']);
 }
